# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Sona API
  version: 0.5.0
servers:
  - url: http://sosee.eu.cc:6699/api/v1
paths:
...
  /me/favorites:
    get:
      operationId: listFavorites
      responses:
        "200":
          description: 当前用户收藏的曲目 ID
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Favorites"
    delete:
      operationId: removeFavorites
      description: 批量取消当前用户收藏，最多 500 首
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [trackIds]
              properties:
                trackIds:
                  type: array
                  minItems: 1
                  maxItems: 500
                  items: {type: string}
      responses:
        "204": {description: 已批量取消收藏}
  /me/favorites/{trackId}:
    put:
      operationId: addFavorite
      parameters:
        - $ref: "#/components/parameters/TrackId"
      responses:
        "204":
          description: 已收藏
    delete:
      operationId: removeFavorite
      parameters:
        - $ref: "#/components/parameters/TrackId"
      responses:
        "204":
          description: 已取消收藏
  /me/favorites/tracks:
    get:
      operationId: favoriteTracks
      parameters:
        - name: cursor
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
      responses:
        "200":
          description: 按收藏时间倒序返回当前用户的收藏歌曲
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrackPage"
  /me/playlists:
    get:
      operationId: listPlaylists
      responses:
        "200":
          description: 当前用户歌单
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Playlist"
    post:
      operationId: createPlaylist
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                  maxLength: 80
      responses:
        "201":
          description: 歌单已创建
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Playlist"
  /me/playlists/{playlistId}:
    delete:
      operationId: deletePlaylist
      parameters:
        - $ref: "#/components/parameters/PlaylistId"
      responses:
        "204":
          description: 歌单已删除
  /me/playlists/{playlistId}/tracks/{trackId}:
    put:
      operationId: addPlaylistTrack
      parameters:
        - $ref: "#/components/parameters/PlaylistId"
        - $ref: "#/components/parameters/TrackId"
      responses:
        "204":
          description: 曲目已加入歌单
    delete:
      operationId: removePlaylistTrack
      parameters:
        - $ref: "#/components/parameters/PlaylistId"
        - $ref: "#/components/parameters/TrackId"
      responses:
        "204":
          description: 曲目已移出歌单
  /me/playlists/{playlistId}/tracks:
    delete:
      operationId: removePlaylistTracks
      description: 批量移出当前用户歌单中的歌曲，最多 500 首
      parameters:
        - $ref: "#/components/parameters/PlaylistId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [trackIds]
              properties:
                trackIds:
                  type: array
                  minItems: 1
                  maxItems: 500
                  items: {type: string}
      responses:
        "204": {description: 已批量移出歌单}
...
  /me/history/{trackId}:
    post:
      operationId: recordPlayback
      parameters:
        - $ref: "#/components/parameters/TrackId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [listenedMs, progressPercent]
              properties:
                listenedMs: {type: integer, minimum: 0}
                progressPercent: {type: number, minimum: 0, maximum: 100}
      responses:
        "204":
          description: 播放记录已保存；发现池歌曲最近 10 次有效播放的平均完播率高于 80% 时自动转入正常池
  /me/tracks/{trackId}:
    delete:
      operationId: hideTrack
      parameters:
        - $ref: "#/components/parameters/TrackId"
      responses:
        "204":
          description: 歌曲已加入当前用户垃圾桶
...
  /me/trash/{trackId}:
    put:
      operationId: restoreTrack
      parameters:
        - $ref: "#/components/parameters/TrackId"
      responses:
        "204": {description: 歌曲已恢复}
  /me/playback-state:
    get:
      operationId: playbackState
      responses:
        "200": {description: 当前用户完整播放队列与进度}
        "204": {description: 暂无续播状态}
    put:
      operationId: savePlaybackState
      responses:
        "204": {description: 续播状态已保存}
...
  /tracks/{trackId}:
    get:
      operationId: getTrack
      parameters:
        - $ref: "#/components/parameters/TrackId"
      responses:
        "200":
          description: 曲目详情
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Track"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  parameters:
    TrackId:
      name: trackId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    PlaylistId:
      name: playlistId
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
...
    NotFound:
      description: 资源不存在
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
  schemas:
...
    Favorites:
      type: object
      required: [trackIds]
      properties:
        trackIds:
          type: array
          items:
            type: string
    Playlist:
      type: object
      required: [id, name, trackIds, createdAt]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        trackIds:
          type: array
          items:
            type: string
        createdAt:
          type: integer
          format: int64
...
    Track:
      type: object
      required: [id, title, artist, album, durationMs, codec, fileExtension, streamURL, hasLyrics, metadataStatus, poolType, audienceType, genre, region, artists]
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        artist:
          type: string
        album:
          type: string
        trackNumber:
          type: integer
        durationMs:
          type: integer
          format: int64
        codec:
          type: string
        fileExtension:
          type: string
        sampleRate:
          type: integer
        bitDepth:
          type: integer
        artworkURL:
          type: string
          format: uri-reference
        streamURL:
          type: string
          format: uri-reference
        hasLyrics:
          type: boolean
        metadataStatus:
          type: string
          enum: [LOCAL, SCRAPED, NEEDS_REVIEW]
        poolType:
          type: string
          enum: [PENDING, NORMAL, DISCOVERY]
        audienceType:
          type: string
          enum: [GENERAL, CHILD]
        genre:
          type: string
        region:
          type: string
          enum: [CN, KR, US, JP, OTHER]
        artists:
          type: array
          description: 艺人归属列表；不切割联名名称，包含“林俊杰”的名称统一归属为“林俊杰”
          items:
            type: string
...
    TrackPage:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Track"
        nextCursor:
          type: [string, "null"]
...
    Problem:
      type: object
      required: [status, title]
      properties:
        status:
          type: integer
        title:
          type: string
        detail:
          type: string